Fixed bad rounding of grades, (used round function from cmath)

+++ CS_Course.h
@@ -14,7 +14,7 @@ private:
     char* Book_Name;
 
 public:
-    CS_Course(int Course_Num, char *Course_Name, int Num_Of_HW, int HW_Weight,bool Is_Takef, char* Book_Name);
+    CS_Course(int Course_Num, char *Course_Name, int Num_Of_HW, double HW_Weight,bool Is_Takef, char* Book_Name);
     ~CS_Course();
     bool isTakef() const;
     char* getBook() const;
